move Kotlin/Native metadata deserialization + KLIB reader to a separate module
