+++ backend.native/compiler/ir/backend.native/src/org/jetbrains/kotlin/backend/konan/serialization/KonanSerializationUtil.kt
@@ -17,32 +17,18 @@
 package org.jetbrains.kotlin.backend.konan.serialization
 
 import org.jetbrains.kotlin.backend.konan.Context
-import org.jetbrains.kotlin.backend.konan.createInteropLibrary
 import org.jetbrains.kotlin.backend.konan.descriptors.isExpectMember
-import org.jetbrains.kotlin.backend.konan.library.KonanLibraryReader
 import org.jetbrains.kotlin.backend.konan.library.LinkData
-import org.jetbrains.kotlin.config.LanguageVersionSettings
-import org.jetbrains.kotlin.descriptors.*
-import org.jetbrains.kotlin.descriptors.impl.ModuleDescriptorImpl
-import org.jetbrains.kotlin.descriptors.konan.DeserializedKonanModuleOrigin
-import org.jetbrains.kotlin.descriptors.konan.createKonanModuleDescriptor
-import org.jetbrains.kotlin.incremental.components.LookupTracker
-import org.jetbrains.kotlin.metadata.konan.KonanProtoBuf
-import org.jetbrains.kotlin.metadata.konan.KonanProtoBuf.*
-import org.jetbrains.kotlin.metadata.ProtoBuf
+import org.jetbrains.kotlin.descriptors.ClassDescriptor
+import org.jetbrains.kotlin.descriptors.DeclarationDescriptor
+import org.jetbrains.kotlin.descriptors.ModuleDescriptor
 import org.jetbrains.kotlin.metadata.deserialization.BinaryVersion
+import org.jetbrains.kotlin.metadata.konan.KonanProtoBuf
 import org.jetbrains.kotlin.name.FqName
-import org.jetbrains.kotlin.name.Name
-import org.jetbrains.kotlin.resolve.CompilerDeserializationConfiguration
 import org.jetbrains.kotlin.resolve.DescriptorUtils
 import org.jetbrains.kotlin.resolve.descriptorUtil.module
 import org.jetbrains.kotlin.resolve.scopes.DescriptorKindFilter
 import org.jetbrains.kotlin.serialization.KonanDescriptorSerializer
-import org.jetbrains.kotlin.serialization.deserialization.*
-import org.jetbrains.kotlin.storage.LockBasedStorageManager
-import org.jetbrains.kotlin.storage.StorageManager
-import org.jetbrains.kotlin.types.KotlinType
-import org.jetbrains.kotlin.types.SimpleType
 
 /*
  * This is Konan specific part of public descriptor 
@@ -54,92 +40,6 @@ import org.jetbrains.kotlin.types.SimpleType
  * The lower level deserializations are performed by the frontend
  * with MemberDeserializer class.
  */
-
-/* ------------ Deserializer part ------------------------------------------*/
-
-object NullFlexibleTypeDeserializer : FlexibleTypeDeserializer {
-    override fun create(proto: ProtoBuf.Type, flexibleId: String, 
-        lowerBound: SimpleType, upperBound: SimpleType): KotlinType {
-            error("Illegal use of flexible type deserializer.")
-        }
-}
-
-fun createKonanPackageFragmentProvider(
-        reader: KonanLibraryReader,
-        fragmentNames: List<String>,
-        storageManager: StorageManager, module: ModuleDescriptor,
-        configuration: DeserializationConfiguration): PackageFragmentProvider {
-
-    val packageFragments = fragmentNames.map{ 
-        KonanPackageFragment(it, reader, storageManager, module)
-    }
-
-    val interopLibrary = createInteropLibrary(reader)
-
-    val syntheticInteropPackageFragments =
-            interopLibrary?.createSyntheticPackages(module, packageFragments) ?: emptyList()
-
-    val provider = PackageFragmentProviderImpl(packageFragments + syntheticInteropPackageFragments)
-
-    val notFoundClasses = NotFoundClasses(storageManager, module)
-
-    val annotationAndConstantLoader = AnnotationAndConstantLoaderImpl(module, notFoundClasses, KonanSerializerProtocol)
-
-    val components = DeserializationComponents(
-        storageManager, module, configuration, 
-        DeserializedClassDataFinder(provider),
-        annotationAndConstantLoader,
-        provider, 
-        LocalClassifierTypeSettings.Default, 
-        ErrorReporter.DO_NOTHING,
-        LookupTracker.DO_NOTHING, NullFlexibleTypeDeserializer,
-        emptyList(), notFoundClasses, ContractDeserializer.DEFAULT, extensionRegistryLite = KonanSerializerProtocol.extensionRegistry )
-
-        for (packageFragment in packageFragments) {
-            packageFragment.initialize(components)
-        }
-
-    return provider
-}
-
-public fun parsePackageFragment(packageData: ByteArray): LinkDataPackageFragment =
-    LinkDataPackageFragment.parseFrom(packageData,
-        KonanSerializerProtocol.extensionRegistry)
-
-public fun parseModuleHeader(libraryData: ByteArray): LinkDataLibrary =
-    LinkDataLibrary.parseFrom(libraryData,
-        KonanSerializerProtocol.extensionRegistry)
-
-public fun emptyPackages(libraryData: ByteArray) 
-    = parseModuleHeader(libraryData).emptyPackageList
-
-internal fun deserializeModule(languageVersionSettings: LanguageVersionSettings,
-                               reader: KonanLibraryReader): ModuleDescriptorImpl {
-
-    val libraryProto = parseModuleHeader(reader.moduleHeaderData)
-
-    val moduleName = libraryProto.moduleName
-
-    val storageManager = LockBasedStorageManager()
-    val moduleDescriptor = createKonanModuleDescriptor(
-            Name.special(moduleName), storageManager,
-            origin = DeserializedKonanModuleOrigin(reader)
-    )
-    val deserializationConfiguration = CompilerDeserializationConfiguration(languageVersionSettings)
-
-    val provider = createKonanPackageFragmentProvider(
-            reader,
-            libraryProto.packageFragmentNameList,
-            storageManager,
-            moduleDescriptor, deserializationConfiguration)
-
-    moduleDescriptor.initialize(provider)
-
-    return moduleDescriptor
-}
-
-
-/* ------------ Serializer part ------------------------------------------*/
 
 internal class KonanSerializationUtil(val context: Context, metadataVersion: BinaryVersion) {
