+++ extracted/konan.serializer/src/org/jetbrains/kotlin/konan/library/KonanLibraryReader.kt
@@ -1,0 +1,30 @@
+package org.jetbrains.kotlin.konan.library
+
+import org.jetbrains.kotlin.config.LanguageVersionSettings
+import org.jetbrains.kotlin.descriptors.ModuleDescriptor
+import org.jetbrains.kotlin.konan.file.File
+import org.jetbrains.kotlin.konan.properties.Properties
+
+interface KonanLibraryReader {
+
+    val libraryFile: File
+    val libraryName: String
+
+    val manifestProperties: Properties
+    val uniqueName: String
+    val linkerOpts: List<String>
+    val unresolvedDependencies: List<String>
+    val bitcodePaths: List<String>
+    val includedPaths: List<String>
+
+    val dataFlowGraph: ByteArray?
+    val moduleHeaderData: ByteArray
+    fun packageMetadata(fqName: String): ByteArray
+
+    val isDefaultLibrary: Boolean get() = false
+    val isNeededForLink: Boolean get() = true
+    val resolvedDependencies: MutableList<KonanLibraryReader>
+    fun markPackageAccessed(fqName: String)
+
+    fun moduleDescriptor(specifics: LanguageVersionSettings): ModuleDescriptor
+}
