+++ extracted/konan.serializer/src/org/jetbrains/kotlin/konan/library/impl/DefaultMetadataReaderImpl.kt
@@ -1,0 +1,12 @@
+package org.jetbrains.kotlin.konan.library.impl
+
+import org.jetbrains.kotlin.konan.library.KonanLibrary
+import org.jetbrains.kotlin.konan.library.MetadataReader
+
+object DefaultMetadataReaderImpl : MetadataReader {
+
+    override fun loadSerializedModule(library: KonanLibrary): ByteArray = library.moduleHeaderFile.readBytes()
+
+    override fun loadSerializedPackageFragment(library: KonanLibrary, fqName: String): ByteArray =
+            library.packageFile(fqName).readBytes()
+}
