+++ extracted/konan.serializer/src/org/jetbrains/kotlin/serialization/konan/interop/InteropDescriptors.kt
@@ -1,22 +1,5 @@
-/*
- * Copyright 2010-2017 JetBrains s.r.o.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
+package org.jetbrains.kotlin.serialization.konan.interop
 
-package org.jetbrains.kotlin.backend.konan.descriptors
-
-import org.jetbrains.kotlin.backend.konan.serialization.KonanPackageFragment
 import org.jetbrains.kotlin.descriptors.ClassifierDescriptor
 import org.jetbrains.kotlin.descriptors.ModuleDescriptor
 import org.jetbrains.kotlin.descriptors.impl.PackageFragmentDescriptorImpl
@@ -25,6 +8,7 @@ import org.jetbrains.kotlin.name.FqName
 import org.jetbrains.kotlin.name.Name
 import org.jetbrains.kotlin.resolve.scopes.MemberScope
 import org.jetbrains.kotlin.resolve.scopes.MemberScopeImpl
+import org.jetbrains.kotlin.serialization.konan.KonanPackageFragment
 import org.jetbrains.kotlin.utils.Printer
 import org.jetbrains.kotlin.utils.addToStdlib.firstNotNullResult
 
@@ -33,7 +17,9 @@ import org.jetbrains.kotlin.utils.addToStdlib.firstNotNullResult
  * redirect "$pkg.$name" to e.g. "cnames.structs.$name".
  */
 class ExportedForwardDeclarationsPackageFragmentDescriptor(
-        module: ModuleDescriptor, fqName: FqName, declarations: List<FqName>
+        module: ModuleDescriptor,
+        fqName: FqName,
+        declarations: List<FqName>
 ) : PackageFragmentDescriptorImpl(module, fqName) {
 
     private val memberScope = object : MemberScopeImpl() {
@@ -66,7 +52,9 @@ class ExportedForwardDeclarationsPackageFragmentDescriptor(
  * The package fragment that redirects all requests for classifier lookup to its targets.
  */
 class ClassifierAliasingPackageFragmentDescriptor(
-        targets: List<KonanPackageFragment>, module: ModuleDescriptor, fqName: FqName
+        targets: List<KonanPackageFragment>,
+        module: ModuleDescriptor,
+        fqName: FqName
 ) : PackageFragmentDescriptorImpl(module, fqName) {
 
     private val memberScope = object : MemberScopeImpl() {
@@ -84,7 +72,7 @@ class ClassifierAliasingPackageFragmentDescriptor(
             p.println(this::class.java.simpleName, " {")
             p.pushIndent()
 
-            p.println("targets = " + targets)
+            p.println("targets = $targets")
 
             p.popIndent()
             p.println("}")
