+++ settings.gradle
@@ -35,7 +35,7 @@ include ':platformLibs'
 includeBuild 'shared'
 includeBuild 'extracted/konan.descriptors'
 includeBuild 'extracted/konan.metadata'
-includeBuild 'extracted/konan.metadata.serializer'
+includeBuild 'extracted/konan.serializer'
 includeBuild 'tools/kotlin-native-gradle-plugin'
 
 if (hasProperty("kotlinProjectPath")) {
