+++ shared/src/main/kotlin/org/jetbrains/kotlin/konan/file/File.kt
@@ -37,6 +37,8 @@ data class File constructor(internal val javaPath: Path) {
         get() = File(absolutePath)
     val name: String
         get() = javaPath.fileName.toString()
+    val extension: String
+        get() = name.substringAfterLast('.', "")
     val parent: String
         get() = javaPath.parent.toString()
     val parentFile: File
